DodatDispley

+++ components/arduino/libraries/WebServer/examples/HelloServer/HelloServer.ino
@@ -1,11 +1,10 @@
-#include <Arduino.h>
 #include <WiFi.h>
-#include <NetworkClient.h>
+#include <WiFiClient.h>
 #include <WebServer.h>
 #include <ESPmDNS.h>
 
-const char *ssid = "........";
-const char *password = "........";
+const char* ssid = "........";
+const char* password = "........";
 
 WebServer server(80);
 
@@ -27,7 +26,7 @@ void handleNotFound() {
   message += "\nArguments: ";
   message += server.args();
   message += "\n";
-  for (int i = 0; i < server.args(); i++) {
+  for (uint8_t i = 0; i < server.args(); i++) {
     message += " " + server.argName(i) + ": " + server.arg(i) + "\n";
   }
   server.send(404, "text/plain", message);
@@ -71,5 +70,5 @@ void setup(void) {
 
 void loop(void) {
   server.handleClient();
-  delay(2);  //allow the cpu to switch to other tasks
+  delay(2);//allow the cpu to switch to other tasks
 }
